feat | refactor: created a new (and unique) button in the home page to redirect the user to main page which will display the charts and other information, also added a new page to show the traffic data in real time like a JSON format.
